Fixed build script

+++ jvm/build.rs
@@ -1,10 +1,15 @@
-use std::process::{Command, Stdio};
+use std::{
+    path::Path,
+    process::{Command, Stdio},
+};
 
 fn main() {
-    println!("cargo:rerun-if-changed=java/com/");
+    println!("cargo:rerun-if-changed=../java/com/");
 
-    println!("Clearing build dir");
-    std::fs::remove_dir_all("../java/out").unwrap();
+    if Path::new("../java/out").exists() {
+        eprintln!("Clearing build dir");
+        std::fs::remove_dir_all("../java/out").unwrap();
+    }
 
     let java_child = Command::new("javac")
         .args([
